moved to same component folder
will implement a fetchington later
